아이템 필터링 관련 커스텀 훅 생성
- ItemListPage, BookmarkListPage 에서 공통으로 사용되는 필터링 로직 커스텀 훅으로 분리
